tests(izone): move discovery unit tests to test_discovery.py
Split discovery service tests out of test_config_flow.py with no
behaviour changes to keep the config flow test file focused.
